[Camera and StaticMesh]
Created a camera that uses a transform component to move about.
The rotation seems to be a bit awkward and not moving according to expectation. Needs an FPS demo made to see whether the view needs inversion or it is the transform multiplication order.

Created StaticMesh class that is the basis for holding mesh data that is ready to be drawn by OGL.
Added helper functions to construct different meshes by initialising and giving the vertex data to a staticmesh.