[Orientation Bug fixed]
So the quaternions are finally behaving correctly. The transform's lookat function was using the glm::lookAt function, which preps a lookat matrix for the VIEW matrix.
This means, the lookat rotation matrix it generates is pre-inverted. To get the original orientation it needed to be conjugated.

The better solution is to not rely on that function and just create a lookAt function for the quaternion class. That way the camera lookat and a transform lookat can be seperate functions.