[Input Singleton + Scene Class + Component Class]
(Input Singleton)
Created an Input singleton to be used to process inputs and allow for key event polling.

Discovered issues with returning a reference to the instance and storing it in a local object variable.
It will create a copy of the object, then it will call its destructor once out of scope.
If this approach is to be repeated, more study is needed to figure out how to stop the reference from being an object copy, and rather an exact reference.
For now it returns a pointer to the instance.

(Scene Class)
Created a Scene class which is used to act as a container for game scenes.

An FPSDemo is used as an example of how to use it to make different game demonstrations.
However functionlity is unclear for now, mainly gameobject component references and how they're added into a scene.
This needs more monitoring as the project continues.

(Component Class)
Added barebones class for component. Not yet used but once gameobject/component updating design has been decided, it will be used.
Currently considering lookup tables for the gameobjects and their components.
i.e. there is a master array of all gameObjects, their position is their ID. Components are all added into a large array and have a reference to their owner ID.
retrieving the owner is just grabbing the struct from the master array.

(Extra)
Fixed rendering of XZ plane for the PrimitiveShapes, the vertex positions were wrong.