[Basic GameObject & Component Implementation]
Made the basic implementation of components and gameobjects. Tried to replicate the component architecture of Unity. Moved the FPSDemo logic into a component that is added to a GameObject.
Further moving down the path of Unity's component architecture schema, needs to include helper functions to initialise and setup major entities.
Also, Component will need to be modified so "new" cannot be used with it except within the GameObject class(friend class?).