[RenderComponent, Friend classes, and Factories]
Created a render component to allow a gameobject to have a mesh attached to it.
Rendercomponent still needs to have some logic to auto-add itself to a render queue.
A render queue is also needed next.

Friend classes are working for some of the logic, but for user made components inherited from the component class, its causing some issues with the constructors.
Derived classes do not inherit the friendship from component, so a factory method will be needed next. Maybe via reflection if possible, or some other method.
Easy solution is to parametise the addcomponent function instead of template it.

Cleaned up unnecessary functions out from other classes and test files.

Camera now has a MainCamera static member that defaults to the first camera to be created when there isn't a camera available.