[GameObject and Component destruction + Toggling]
Added ability to toggle the gameobject's active state which affects all its children components.
Added ability to destroy a gameobject which destroys itself and cleans up the components too.

Noted down ideas for how a factory design would work for the objects in this framework in the to-do list.
Also learned Component's destructor was not virtual, so it was causing the derived classes to not have their destructors called.