[Material class]
Simple material class added to consolidate a shader with a texture together.
Needs to be expanded a bit to allow for different shaders and texture combinations.
A shader compiler class or material compiler class needs to be made to make sure shaders aren't compiled again and again.
Also to share common shaders amongst materials. Or a material class that manages all active materials and cleans up after them.