[Object Factory Redesign]
Redesigned gameobject and component to have a base Object class, that handles instantiation and memory allocation.
The Instantiate and Destroy function for GameObject was modified to reflect this. Also, the Destroy function now clears the pointer it is passed.
Shared_ptr from the standard library would be better suited, or perhaps redesigning so 'new' and 'delete' can be valid keywords to help readability.
The new and delete keywords could be overwritten to return a shared_ptr reference, so that delete removes it from the major list, and new creates an instance and keeps only one.