[GameObject - Self-cleaning references]
Implemented a container interface for the GameObject class to keep 1 unique instance of an Instantiated GameObject.
It's implementation was tested in an isolated project from this one, and integrated over. Thorough testing is still required, but at this time it functions as expected.

The GameObject keeps an instance of itself when its instantiated, and returns a reference to itself via operator overloading the pointer operators.
The '.' operator will need to be blocked so only '->' can be used with the class. It will semantically confuse things though, so a more cleaner interface needs to be designed.