[Text Rendering Component]
Began the makings of a UIRenderComponent that for now just renders text.
Spent way too long trying to get this to work with triangle strips... winding order and uvs were too pesky, so regular triangle list is used.
Based on the learnopengl.com text rendering tutorial. The tutorial has a few winding issues in the render plane for the text ui.

Next, more debugging facilities to visualise the parameters and data of entities.
-Render their orientation
-Create bounding boxes and render them

Also, it is currently a bit cumbersome to create a simple mesh and render it in the framework.
A simple helper function to setup a ready gameobject or a more intuitive object that neglects the render pipeline and can be used outside of the framework for debugging purposes is needed.
Adding an "opengl code" object, which allows the user to inject their own opengl code to be run independently might be useful... but it defeats the purpose of using the framework then.
As this is using C++ and opengl it is possible to just write opengl code anywhere, but offering a VertexBufferObject/Array object that can have data and buffers prepared easily might be useful.