[Transform Visualiser Component + Dynamic Mesh]
Created a Dynamic Mesh class better suited to meshes that need to be changed frequently.
Refactoring will be needed for the rendercomponent to better deal with taking a staticmesh or a dynamic mesh.

A Transform Visualiser Component was created to visualise and show the attached game object's transform orientation and origin position.
This took a week to debug, due to not properly understanding VAOs, VBOs and EBOs... Also the initial setup of the BufferData was not large enough in memory allocation to store the sub buffer data calls.

After this, all code relating to rendering needs to be checked and refactored to reflect proper understanding of OpenGL's state machine objects.

A bug was caught in the Input class that didn't account for an unknown key (which returns -1 as its code value). This needs to be better explored later.