[DynamicMesh Refactoring & To-do update]
Refactored DynamicMesh to take into account the updated knowledge about VAOs in OpenGL.
Added a dirty flag for DynamicMesh to let it know if it needs to be updated and let it update the proper data accordingly for the buffers.
Changed the draw function for the transform visualiser to consolidate the axis draw calls.

To-do list updated to reflect next tasks and comment on those that have been done or deprecated.