[Vertex Array Object class]
Began refactoring process for the rendering of meshes and other primitives.
Created a class to handle creation of buffers and setting up the vertex attributes.
Needs to have the ability to let the buffer be easily modifiable.
Should also keep track of the current VBO and EBO, while allowing for multiple buffers and having an interface to let the user choose a buffer, and modify it.
Error handling for the buffer's size should be checked when data copy is requested and glBufferData should be called again to allow for the correction.