[Vertex Array Object + Mesh Decoupling]
Finally have a working example in main of them being decoupled.
However there is still a little bit of a cumbersome interaction between the 2 of them.
For now it's working, the next stage is to refactor "StaticMesh" and "DynamicMesh"  either not integrate the VAO and leave that to the upcoming MeshRenderComponent, or integrate a VAO with the mesh and let them handle their own VAOs.
We'll see after attempting refactor.