[MeshRendererComponent]
Added a mesh renderer component to finally decouple the 'mesh' from a rendercomponent, however a new issue has been found.
The VAO class is expecting to know mesh data when it is setting up the buffers, which is fine, but then when a user modifies or changes the mesh, there is no way to update the buffer properties when the change occurs.
A fiddly way can be done by calling the copyToBuffer function in the VAO but this is using it for an unintended purpose and doesn't communicate clear interfacing with the object.

A new way to update the buffer properties needs to be added.
An extra thing for consideration is, considering the Mesh class already handles allowing itself to be modified, it makes sense for it to store mesh buffer property data.
As for the material, a Mesh could hold it's own material, but for now keeping the material within the MeshRenderer will allow for it to handle setting up the VertexAttributes on a per material basis.
Which brings up a new consideration point, should a Shader/Material create or require VertexAttribute data that can be used by the VAO to set itself up. Thoughts for next time.