[Material + VertexAttributes integration]
Integrated the inclusion of VertexAttributes in the Material, so now when creating a material it can have it's attributes prepared and when creating a mesh it can have its buffer prepared.
Then a VAO can take the data from both these things and create a VertexArray for OpenGL to use for rendering.
Found a bug where using set mesh in MeshRenderComponent wouldn't trigger the update flags to update in mesh, as a copy constructor is called so the flags are never set.

(P.S. Taking a month break from here sure was giving me a headscratcher when trying to follow how my code was working :P)